Refactor TS metadata_util into parts in Galaxy/TS.
